last udate 11:35 pm fri 21/7

+++ back-end/controller/userController.js
@@ -91,7 +91,9 @@ async function deleteU(req, res) {
       return res.status(404).json({ errors: ["User not found"] });
     }
 
-    fs.unlinkSync("./upload/" + user[0].image);
+    if(user[0].image){
+      fs.unlinkSync("./upload/" + user[0].image);
+      }
 
     await deleteUser(user[0].id);
